+++ src/daemon/error.rs
