--- src/tray_daemon/frontend/mod.rs
@@ -1,8 +1,0 @@
-pub mod file_chooser;
-pub mod main_window;
-pub mod screenshot_window;
-pub mod ui;
-pub mod ui_manager;
-
-#[derive(thiserror::Error, Debug)]
-pub enum Error {}
